Functionality 2
I used a basic list structure to enhance Functionality 1, which I developed in Week 1. I allowed users to enter information for five employees simultaneously, where the first employee information entry is saved in index [0] the second entry is saved in index [1], and so on. I made sure to enforce the user to enter the following employee information: employeeName, employeeSSN, employeePhone, employeeEmail, employeeSalary. After the user finishes entering the last employee information, they should get a message to allow them to print specific employee information from the saved list. The user can enter values between 1 and 5, where 1 will return the employee information saved in the list index [0], 2 will return the employee information saved in the list index [1] and so on. The printed result should be in the following format:

Mike Smith, 123123,(111)222-3333,mike@gmail.com,$6000