Add files via upload
In this functionality, I utilized looping to make my Python script for Functionality 2  to run constantly. I then proceeded to separate Functionality 1. into the following two functionalities:
Add Employee – this functionality will allow users to add new employee to the system.
View all Employees – this functionality will view all employees in the system.
Use global variables to develop a counter to keep track of the number of employees in the system. A good employee management system should always give brief information about the existing number of employees. This counter can be shown to the user when they run the script as the following message:
There are (3) employees in the system.